+++ lib/types/core/base64.rb
@@ -1,0 +1,8 @@
+rdl_nowrap :Base64
+
+type :Base64, 'self.decode64', '(String) -> String'
+type :Base64, 'self.encode64', '(String) -> String'
+type :Base64, 'self.strict_decode64', '(String) -> String'
+type :Base64, 'self.strict_encode64', '(String) -> String'
+type :Base64, 'self.urlsafe_decode64', '(String) -> String'
+type :Base64, 'self.urlsafe_encode64', '(String) -> String'