+++ CHANGES.md
@@ -17,6 +17,7 @@
 - Support for instantiate! for binding type parameters during static type checking
 - New "check" flag for calls to instantiate! indicating whether we want to check type of receiving object on call
 - More precise static type checking for `Object#class` method
+- Klass argument to rdl_nowrap, rdl_alias
 
 ### Fixed
 - Parsing bug with `or`